+++ packages/nimble-components/src/rich-text/editor/testing/rich-text-editor.pageobject.ts
@@ -55,6 +55,18 @@ export class RichTextEditorPageObject {
         await waitForUpdatesAsync();
     }
 
+    public async pressShiftEnterKeysInEditor(): Promise<void> {
+        const editor = this.getTiptapEditor();
+        const shiftEnterEvent = new KeyboardEvent('keydown', {
+            key: keyEnter,
+            shiftKey: true,
+            bubbles: true,
+            cancelable: true
+        });
+        editor!.dispatchEvent(shiftEnterEvent);
+        await waitForUpdatesAsync();
+    }
+
     public async pressTabKeyInEditor(): Promise<void> {
         const editor = this.getTiptapEditor();
         const event = new KeyboardEvent('keydown', {
@@ -117,6 +129,13 @@ export class RichTextEditorPageObject {
     }
 
     public async setEditorTextContent(value: string): Promise<void> {
+        const lastElement = this.getEditorLastChildElement();
+        const textNode = document.createTextNode(value);
+        lastElement.parentElement!.appendChild(textNode);
+        await waitForUpdatesAsync();
+    }
+
+    public async replaceEditorContent(value: string): Promise<void> {
         const lastElement = this.getEditorLastChildElement();
         lastElement.parentElement!.textContent = value;
         await waitForUpdatesAsync();