+++ packages/nimble-components/src/rich-text/models/tests/markdown-serializer.spec.ts
@@ -2,6 +2,7 @@ import { Editor } from '@tiptap/core';
 import Bold from '@tiptap/extension-bold';
 import BulletList from '@tiptap/extension-bullet-list';
 import Document from '@tiptap/extension-document';
+import HardBreak from '@tiptap/extension-hard-break';
 import Italic from '@tiptap/extension-italic';
 import ListItem from '@tiptap/extension-list-item';
 import OrderedList from '@tiptap/extension-ordered-list';
@@ -24,6 +25,7 @@ describe('Markdown serializer', () => {
             ListItem,
             Bold,
             Italic,
+            HardBreak,
             Link.extend({
                 excludes: '_'
             })
@@ -272,11 +274,6 @@ describe('Markdown serializer', () => {
             },
             { name: 'Heading', html: '<h1>Heading</h1>', plainText: 'Heading' },
             {
-                name: 'HardBreak',
-                html: '<p>Hard<br>Break</p>',
-                plainText: 'Hard Break'
-            },
-            {
                 name: 'HorizontalRule',
                 html: '<p>Horizontal<hr>Rule</p>',
                 plainText: 'Horizontal\n\nRule'
@@ -327,6 +324,90 @@ describe('Markdown serializer', () => {
                     expect(
                         RichTextMarkdownSerializer.serializeDOMToMarkdown(node)
                     ).toBe(value.plainText);
+                }
+            );
+        }
+    });
+
+    describe('HardBreak node should be serialized to back slash (hard break syntax) markdown output', () => {
+        const r = String.raw;
+        const supportedNodesMarks: {
+            name: string,
+            html: string,
+            markdown: string
+        }[] = [
+            {
+                name: 'Hard Break',
+                html: '<p>Hard<br>Break</p>',
+                markdown: r`Hard\
+Break`
+            },
+            {
+                name: 'Bold',
+                html: '<strong>Bold</strong><br><strong>Bold</strong>',
+                markdown: r`**Bold**\
+**Bold**`
+            },
+            {
+                name: 'Italics',
+                html: '<em>Italics</em><br><em>Italics</em>',
+                markdown: r`*Italics*\
+*Italics*`
+            },
+            {
+                name: 'Bold, Hard break and Italics',
+                html: '<strong>Bold</strong><br><em>Italics</em>',
+                markdown: r`**Bold**\
+*Italics*`
+            },
+            {
+                name: 'Numbered list',
+                html: '<ol><li><p>Numbered<br>list</p></li></ol>',
+                markdown: r`1. Numbered\
+   list`
+            },
+            {
+                name: 'Bulleted list',
+                html: '<ul><li><p>Bulleted<br>list</p></li></ul>',
+                markdown: r`* Bulleted\
+  list`
+            },
+            {
+                name: 'Nested Bulleted list and hard break',
+                html: '<ul><li><p>list<br>hard break content</p></li><li><p>list</p><ul><li><p>nested list<br>nested hard break content</p></li></ul></li></ul>',
+                markdown: r`* list\
+  hard break content
+
+* list
+
+  * nested list\
+    nested hard break content`
+            },
+            {
+                name: 'Nested Numbered list and hard break',
+                html: '<ol><li><p>list<br>hard break content</p></li><li><p>list</p><ol><li><p>nested list<br>nested hard break content</p></li></ol></li></ol>',
+                markdown: r`1. list\
+   hard break content
+
+2. list
+
+   1. nested list\
+      nested hard break content`
+            }
+        ];
+
+        const focused: string[] = [];
+        const disabled: string[] = [];
+        for (const value of supportedNodesMarks) {
+            const specType = getSpecTypeByNamedList(value, focused, disabled);
+            specType(
+                `Should serialize ${value.name} to markdown`,
+                // eslint-disable-next-line @typescript-eslint/no-loop-func
+                () => {
+                    const node = getNode(value.html);
+                    expect(
+                        RichTextMarkdownSerializer.serializeDOMToMarkdown(node)
+                    ).toBe(value.markdown);
                 }
             );
         }